Add radio functionality to the site
This feature adds a new component called AudioPlayer. This is activated
on the Terminal via the radio command. This will play client-side audio
in the form of mp3s included in the public/mp3 directory. There are
three radio stations: a video games music channel, movie music, and
classical music. In most cases, these are mp3 recordings of midi files.