Get Cat component fully working
Previously, the terminal was not scrolling to the bottom when the pictures appeared. This was because we were not waiting for images to load, so it would not have been able to properly scroll to the bottom.

Now, we're tied into the onload callback for the images, so we don't try to scroll until image has loaded.

This component needed access to the input ref, so that was hoisted out of the Terminal into the page, and is now passed into both the Terminal and Cat components as props.